Deduplicate train_p3 and train_p3_centralized (closes #88)
Collapse the two bucket-brigade P3 examples into a single binary that
toggles the centralized-critic path via a `--centralized-critic` CLI
flag. The shared core (env construction, redundancy-penalty aux loss,
training loop, checkpoint I/O) lives in one place; the two modes diverge
only at trainer construction (`JointMultiAgentTrainer::new` vs
`::new_with_centralized_critic`).

- Add `--centralized-critic`, `--critic-hidden-dim`, `--critic-lr`,
  `--critic-vf-coef`, `--critic-clip-range-vf` flags to `train_p3.rs`.
- Branch on `--centralized-critic` at trainer construction; print a
  per-iteration log line that includes the critic stats when active.
- Delete `examples/games/bucket_brigade/train_p3_centralized.rs` and
  remove its `[[example]]` entry from Cargo.toml.
- Update the binary's module docs to describe both modes.

Verified by running both smoke tests from the issue's test plan:
- `--scenario default --seed 42 --num-updates 2` runs the no-critic
  baseline.
- `--scenario default --seed 42 --num-updates 2 --centralized-critic`
  produces a positive, finite `centralized_value_loss` (~4.2e6 on
  iteration 0). Smoke runs require the `env-bucket-brigade` feature,
  which is gated off in the published manifest (see Cargo.toml notes
  re: `bucket-brigade-core` path dep); tests were run against a locally
  uncommented copy of the feature, which was reverted before committing.
- `cargo test --features training` passes (216 lib + integration tests,
  0 failed).
- `cargo publish --dry-run` still succeeds with the feature gated off.

Net delta: -234 LOC (122 added, 356 deleted).
